Module arm7tdmi_aeabi::fn_declarations

This holds all the extern "C" declarations for the assembly code that the crate can provide.

Functions

aeabi_idiv⚠

Performs i32 / i32 division.

aeabi_idivmod⚠

Performs [i32 / i32, i32 % i32], returning the data packed in a u64.

aeabi_memclr⚠

Sets n bytes in dest to 0.

aeabi_memclr4⚠

Sets n bytes in dest to 0.

aeabi_memclr8⚠

Sets n bytes in dest to 0.

aeabi_memcpy⚠

Copies n bytes to dest from src.

aeabi_memcpy4⚠

Copies n bytes to dest from src.

aeabi_memcpy8⚠

Copies n bytes to dest from src.

aeabi_memmove⚠

Copies n bytes to dest from src (overlap allowed).

aeabi_memmove4⚠

Copies n bytes to dest from src (overlap allowed).

aeabi_memmove8⚠

Copies n bytes to dest from src (overlap allowed).

aeabi_memset⚠

Sets n bytes in dest to val as u8.

aeabi_memset4⚠

Sets n bytes in dest to val as u8.

aeabi_memset8⚠

Sets n bytes in dest to val as u8.

aeabi_uidiv⚠

Performs u32 / u32 division.

aeabi_uidivmod⚠

Performs [u32 / u32, u32 % u32], returning the data packed in a u64.

aeabi_uread4⚠

Safely performs an unaligned read of the u32 address given.

aeabi_uread8⚠

Safely performs an unaligned read of the u64 address given.

aeabi_uwrite4⚠

Safely performs an unaligned write of the u32 address given.

aeabi_uwrite8⚠

Safely performs an unaligned write of the u64 address given.

gba_memcpy_sram⚠

Copies n bytes to dest from src.

libc_memcpy⚠

Copies n bytes to dest from src.

libc_memmove⚠

Copies n bytes to dest from src (overlap allowed).

libc_memset⚠

Sets n bytes in dest to val as u8.